shoppingbag & sold item cal fix

+++ src/app/index/item-sold/item-sold.page.ts
@@ -50,7 +50,7 @@ export class ItemSoldPage implements OnInit {
       this.finaltotal = 0;
       this.item.forEach((item) => {
         this.number1 += 1;
-        this.totalcost += parseInt(item.price);
+        this.totalcost += parseFloat(item.price);
       });
       this.finaltotal = (this.totalcost + this.shipping)/ this.discount;
     })
